Support discovering properties via reflection
Initially I thought from looking at documentation that many of the Reflection interfaces which were needed for this were not supported in Portable Class Libraries. However, it turned out that the Portable Class Library icon does not appear with the method documentation, but at the class level.

So far this is only tested with ViewModel.Number. Code needs to be added to handle Property.SubProperty through reflection and to replace the IsGettable and IsSettable properties with the standard CanWrite and CanRead properties.