Test that binding a source without a getter fails
This is a design change. While an ungettable property could be a binding source in theory, it really does not make much sense. An ungettable property is little more than a command or trigger with a value. So, based upon what a binding user would really want to do this change was made.