Remove null getter and setter exceptions
MakeSmartSetter is only called for fields and when the getter and setter have already been verified to be non-null, so there is no need for it to test them and possibly throw an exception. Also, since MakeNotifyingSetter is only ever called by MakeSmartSetter, its checks can be removed for the same reason.