Simplify PropertyChanged notification
There are now SetProperty and GetProperty extension methods which take property expressions like () => Color to specify the name of the property instead of having to type the name in a string and risking typos.

Also the SetProperty method takes a setting function which is only called if the value has changed. SetProperty also automatically handles sending the PropertyChanged notification using the RaisePropertyChanged method, which has been added to the INotifyingObject interface.

INotifyingObject.OnPropertyChangedEvent has also been deprecated. I am following the model of using Raise... functions to fire events and On... functions to handle them, like OnClicked.