Hoist Property.Registry into Property itself
The static nested Property.Registry class causes some interesting issues in applying ReSharper cleanups to it and the Property class. Since it is really just an encapsulation of a static field and the related methods, all of which are only accessed by Property, all of it can be lifted directly into Property itself with a little renaming.